+++ index.d.ts
@@ -1,5 +1,5 @@
-export {DataTable, DataEvent, PageEvent, SortEvent} from './lib/DataTable';
-export {DefaultSorter} from './lib/DefaultSorter';
-export {Paginator} from './lib/Paginator';
-export {BootstrapPaginator} from './lib/BootstrapPaginator';
-export {DataTableModule} from './lib/DataTableModule';
+export {ServerDataTable, ServerDataEvent, ServerPageEvent, ServerSortEvent} from './lib/ServerDataTable';
+export {ServerDefaultSorter} from './lib/ServerDefaultSorter';
+export {ServerPaginator} from './lib/ServerPaginator';
+export {ServerBootstrapPaginator} from './lib/ServerBootstrapPaginator';
+export {ServerDataTableModule} from './lib/ServerDataTableModule';