+++ src/ServerDefaultSorter.ts
@@ -1,8 +1,8 @@
 import {Component, Input, OnInit} from "@angular/core";
-import {DataTable, SortEvent} from "./DataTable";
+import {ServerDataTable, ServerSortEvent} from "./ServerDataTable";
 
 @Component({
-    selector: "mfDefaultSorter",
+    selector: "smfDefaultSorter",
     template: `
         <a style="cursor: pointer" (click)="sort()" class="text-nowrap">
             <ng-content></ng-content>
@@ -10,17 +10,17 @@ import {DataTable, SortEvent} from "./DataTable";
             <span *ngIf="isSortedByMeDesc" class="glyphicon glyphicon-triangle-bottom" aria-hidden="true"></span>
         </a>`
 })
-export class DefaultSorter implements OnInit {
+export class ServerDefaultSorter implements OnInit {
     @Input("by") sortBy: string;
 
     isSortedByMeAsc: boolean = false;
     isSortedByMeDesc: boolean = false;
 
-    public constructor(private mfTable: DataTable) {
+    public constructor(private smfTable: ServerDataTable) {
     }
 
     public ngOnInit(): void {
-        this.mfTable.onSortChange.subscribe((event: SortEvent) => {
+        this.smfTable.onSortChange.subscribe((event: ServerSortEvent) => {
             this.isSortedByMeAsc = (event.sortBy == this.sortBy && event.sortOrder == "asc");
             this.isSortedByMeDesc = (event.sortBy == this.sortBy && event.sortOrder == "desc");
         });
@@ -28,9 +28,9 @@ export class DefaultSorter implements OnInit {
 
     sort() {
         if (this.isSortedByMeAsc) {
-            this.mfTable.setSort(this.sortBy, "desc");
+            this.smfTable.setSort(this.sortBy, "desc");
         } else {
-            this.mfTable.setSort(this.sortBy, "asc");
+            this.smfTable.setSort(this.sortBy, "asc");
         }
     }
 }